+++ src/array/ft_array_new.c
@@ -1,16 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_array.c                                         :+:      :+:    :+:   */
+/*   ft_array_new.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pqueiroz <pqueiroz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/05/08 21:17:58 by pqueiroz          #+#    #+#             */
-/*   Updated: 2019/05/08 22:59:53 by pqueiroz         ###   ########.fr       */
+/*   Created: 2019/07/12 17:27:18 by pqueiroz          #+#    #+#             */
+/*   Updated: 2019/07/12 17:29:35 by pqueiroz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft.h"
+#include "ft_array.h"
+
 
 t_array	*ft_array_new(void)
 {
@@ -23,30 +24,29 @@ t_array	*ft_array_new(void)
 	return (self);
 }
 
-t_array	*ft_array_new_with_values(void **values)
+/*
+** Instantiate a `t_array *` from an array of void pointers and a length `n`. If
+** length `n` is -1 the array `values` must be null-terminated.
+*/
+
+t_array	*ft_array_new_with_values(void **values, ssize_t n)
 {
 	t_array *self;
 
 	self = ft_array_new();
-	while (values && *values)
-		ft_array_push(self, *values++);
+	if (n == -1)
+	{
+		while (values && *values)
+			ft_array_push(self, *values++);
+	}
+	else
+	{
+		self->size = ((n / ARRAY_GROW_SIZE) + 1) * ARRAY_GROW_SIZE;
+		self->data = malloc(sizeof(*self->data) * self->size);
+		self->length = n;
+		while (--n >= 0)
+			self->data[n] = values[n];
+		self->data[self->length] = NULL;
+	}
 	return (self);
-}
-
-void	ft_array_destroy(t_array **self)
-{
-	if ((*self)->data)
-		free((*self)->data);
-	free(*self);
-	*self = NULL;
-}
-
-void	ft_array_terminate(t_array **self)
-{
-	size_t i;
-
-	i = 0;
-	while (i < (*self)->length)
-		ft_memdel((*self)->data + i++);
-	ft_array_destroy(self);
 }